Optimise casts for clear error messages

+++ inst/library.c
@@ -360,7 +360,7 @@ void rmhyper(size_t n_sample, int *k, size_t m) {
   for (size_t i = 0; i < m; ++i) {
     N += k[i];
   }
-  if (n_sample > (size_t)N) {
+  if ((int)n_sample > N) {
     Rf_error("Requesting too many elements in rmhyper (%d from %d)",
              n_sample, N);
   }
